+++ meteor/client/ui/Settings/ShowStyle/BlueprintConfiguration.tsx
@@ -1,18 +1,18 @@
 import React, { useCallback, useMemo } from 'react'
-import { useTracker } from '../../lib/ReactMeteorData/react-meteor-data'
+import { useTracker } from '../../../lib/ReactMeteorData/react-meteor-data'
 import { BlueprintManifestType, JSONSchema } from '@sofie-automation/blueprints-integration'
-import { BlueprintConfigManifestSettings, SourceLayerDropdownOption } from './BlueprintConfigManifest'
+import { BlueprintConfigManifestSettings, SourceLayerDropdownOption } from '../BlueprintConfigManifest'
 import { SomeObjectOverrideOp } from '@sofie-automation/corelib/dist/settings/objectWithOverrides'
-import { Blueprints, ShowStyleBases } from '../../collections'
+import { Blueprints, ShowStyleBases } from '../../../collections'
 import { useTranslation } from 'react-i18next'
 import { MappingsExt } from '@sofie-automation/corelib/dist/dataModel/Studio'
 import { faExclamationTriangle } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { EditAttribute } from '../../lib/EditAttribute'
-import { RedirectToBlueprintButton } from '../../lib/SettingsNavigation'
+import { EditAttribute } from '../../../lib/EditAttribute'
+import { RedirectToBlueprintButton } from '../../../lib/SettingsNavigation'
 import { unprotectString } from '@sofie-automation/corelib/dist/protectedString'
 import { BlueprintId } from '@sofie-automation/corelib/dist/dataModel/Ids'
-import { ShowStyleBase } from '../../../lib/collections/ShowStyleBases'
+import { ShowStyleBase } from '../../../../lib/collections/ShowStyleBases'
 
 interface ShowStyleBaseBlueprintConfigurationSettingsProps {
 	showStyleBase: ShowStyleBase
