+++ meteor/client/ui/Settings/Studio/BlueprintConfiguration.tsx
@@ -1,19 +1,19 @@
 import React, { useCallback, useMemo } from 'react'
-import { useTracker } from '../../lib/ReactMeteorData/react-meteor-data'
+import { useTracker } from '../../../lib/ReactMeteorData/react-meteor-data'
 import { BlueprintManifestType } from '@sofie-automation/blueprints-integration'
-import { BlueprintConfigManifestSettings } from './BlueprintConfigManifest'
+import { BlueprintConfigManifestSettings } from '../BlueprintConfigManifest'
 import {
 	applyAndValidateOverrides,
 	SomeObjectOverrideOp,
 } from '@sofie-automation/corelib/dist/settings/objectWithOverrides'
-import { Blueprints, Studios } from '../../collections'
+import { Blueprints, Studios } from '../../../collections'
 import { JSONBlobParse } from '@sofie-automation/shared-lib/dist/lib/JSONBlob'
 import { useTranslation } from 'react-i18next'
 import { DBStudio } from '@sofie-automation/corelib/dist/dataModel/Studio'
 import { faExclamationTriangle } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { EditAttribute } from '../../lib/EditAttribute'
-import { RedirectToBlueprintButton } from '../../lib/SettingsNavigation'
+import { EditAttribute } from '../../../lib/EditAttribute'
+import { RedirectToBlueprintButton } from '../../../lib/SettingsNavigation'
 import { Blueprint } from '@sofie-automation/corelib/dist/dataModel/Blueprint'
 import { protectString, unprotectString } from '@sofie-automation/corelib/dist/protectedString'
 import { BlueprintId } from '@sofie-automation/corelib/dist/dataModel/Ids'
